+++ packages/server/lib/browsers/chrome.ts
@@ -15,9 +15,9 @@ import { CdpAutomation, screencastOpts } from './cdp_automation'
 import * as protocol from './protocol'
 import utils from './utils'
 import * as errors from '../errors'
-import { BrowserCriClient } from './browser-cri-client'
+import { BrowserCriClient } from './remote-interface/browser-cri-client'
 import type { Browser, BrowserInstance, GracefulShutdownOptions } from './types'
-import type { CriClient } from './cri-client'
+import type { CriClient } from './remote-interface/cri-client'
 import type { Automation } from '../automation'
 import memory from './memory'
