+++ packages/server/lib/browsers/firefox-util.ts
@@ -7,7 +7,7 @@ import util from 'util'
 import Foxdriver from '@benmalka/foxdriver'
 import * as protocol from './protocol'
 import { CdpAutomation } from './cdp_automation'
-import { BrowserCriClient } from './browser-cri-client'
+import { BrowserCriClient } from './remote-interface/browser-cri-client'
 import type { Automation } from '../automation'
 
 const errors = require('../errors')
