+++ packages/server/lib/browsers/firefox.ts
@@ -16,7 +16,7 @@ import os from 'os'
 import treeKill from 'tree-kill'
 import mimeDb from 'mime-db'
 import { getRemoteDebuggingPort } from './protocol'
-import type { BrowserCriClient } from './browser-cri-client'
+import type { BrowserCriClient } from './remote-interface/browser-cri-client'
 import type { Automation } from '../automation'
 import { getCtx } from '@packages/data-context'
 import { getError } from '@packages/errors'
