+++ packages/server/test/integration/cypress_spec.js
@@ -44,7 +44,7 @@ const appData = require(`../../lib/util/app_data`)
 const electronApp = require('../../lib/util/electron-app')
 const savedState = require(`../../lib/saved_state`)
 const { getCtx, clearCtx, setCtx, makeDataContext } = require(`../../lib/makeDataContext`)
-const { BrowserCriClient } = require(`../../lib/browsers/browser-cri-client`)
+const { BrowserCriClient } = require(`../../lib/browsers/remote-interface/browser-cri-client`)
 const { cloudRecommendationMessage } = require('../../lib/util/print-run')
 
 const processVersions = process.versions
