+++ packages/server/test/unit/browsers/firefox_spec.ts
@@ -11,7 +11,7 @@ import Foxdriver from '@benmalka/foxdriver'
 import * as firefox from '../../../lib/browsers/firefox'
 import firefoxUtil from '../../../lib/browsers/firefox-util'
 import { CdpAutomation } from '../../../lib/browsers/cdp_automation'
-import { BrowserCriClient } from '../../../lib/browsers/browser-cri-client'
+import { BrowserCriClient } from '../../../lib/browsers/remote-interface/browser-cri-client'
 import { ICriClient } from '../../../lib/browsers/cri-client'
 
 const path = require('path')
